Add files via upload
 This project will initiate a new area of research in the field of reversible sequential circuit. In 
today’s world, the complexity of the chip is increasing as more and more devices are being 
connected on a single chip. Due to the high density of the chip, the power dissipation 
increases demanding better power optimization methods. One of the methods to achieve 
power optimization is by using reversible logic. It can be used in low power CMOS designs, 
quantum computing, nanotechnology and optical computing. Optimization of the sequential 
circuit is achieved on the basis of total number of gates used in the circuit. The reversible logic
 design attracts more interest due to its low power consumption.